Add mapping function

+++ pytorch_generative/models/pixel_snail.py
@@ -279,7 +279,7 @@ def reproduce(
     scheduler = lr_scheduler.MultiplicativeLR(optimizer, lr_lambda=lambda _: 0.999977)
 
     def loss_fn(x, _, preds):
-
+        x = x.long()
         criterion = nn.CrossEntropyLoss()
         B, C, D = preds.size()
         preds_2d = preds.view(B, C, D, -1)
